First commit
This program is to make a dupeable html element,it turn to be more codely in js,if you really want to make dupe bottun is inside every unit,it just complicate the way to generate new id and check whether that id is already existed.
--
it tooks me 2 days to made...most of time is trying to do something reasonable like why "getElementsById" isn`t a function and why getElementById is not working in clone node despite I had make that id be unique in elements.